feat : EntityFinder 추가
Entity 조회 메서드를 EntityFinder에 모음
