+++ src/main/java/com/goorm/thelastsupper/reservation/service/ReservationService.java
@@ -11,6 +11,7 @@ import com.goorm.thelastsupper.reservation.exception.ReservationErrorCode;
 import com.goorm.thelastsupper.reservation.exception.ReservationException;
 import com.goorm.thelastsupper.reservation.repository.ReservationHistoryRepository;
 import com.goorm.thelastsupper.reservation.repository.ReservationSlotRepository;
+import com.goorm.thelastsupper.reservation.util.EntityFinder;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -26,17 +27,16 @@ import static com.goorm.thelastsupper.reservation.util.ReservationUtil.*;
 @RequiredArgsConstructor
 public class ReservationService {
 
-    private final ReservationSlotRepository reservationSlotRepository;
     private final ReservationHistoryRepository reservationHistoryRepository;
-    private final AccountRepository accountRepository;
+    private final EntityFinder entityFinder;
 
     @Transactional
     public ReservationResponse registerReservation(String accountId, @Valid ReservationRequest request) {
 
-        ReservationSlot reservationSlot = getReservationSlotById(request.slotId());
+        ReservationSlot reservationSlot = entityFinder.getReservationSlotById(request.slotId());
         log.info("예약 슬롯 조회 성공: slotId={}, date={}, remaining={}", reservationSlot.getId(), reservationSlot.getDate(), reservationSlot.getRemaining());
 
-        Account account = getAccountById(accountId);
+        Account account = entityFinder.getAccountById(accountId);
         log.info("계정 조회 성공: accountId={}, email={}", account.getId(), account.getEmail());
 
         log.info("슬롯 오픈 상태 검증 시작...");
@@ -70,11 +70,11 @@ public class ReservationService {
     @Transactional
     public void cancelReservation(String slotId, String historyId, String accountId) {
 
-        ReservationSlot reservationSlot = getReservationSlotById(slotId);
+        ReservationSlot reservationSlot = entityFinder.getReservationSlotById(slotId);
         log.info("예약 슬롯 조회 성공: slotId={}, date={}, remaining={}", reservationSlot.getId(), reservationSlot.getDate(), reservationSlot.getRemaining());
 
         log.info("예약 존재 여부 확인 시작...");
-        ReservationHistory reservationHistory = getHistoryById(historyId);
+        ReservationHistory reservationHistory = entityFinder.getHistoryById(historyId);
         validateReservationStatus(reservationHistory);
 
         log.info("본인 여부 확인 시작...");
@@ -98,7 +98,7 @@ public class ReservationService {
     @Transactional
     public ReservationResponse modifyReservation(String historyId, String accountId, ReservationRequest request) {
 
-        ReservationHistory reservationHistory = getHistoryById(historyId);
+        ReservationHistory reservationHistory = entityFinder.getHistoryById(historyId);
 
         String historyAccountID = reservationHistory.getAccount().getId();
 
@@ -109,30 +109,5 @@ public class ReservationService {
 
         return registerReservation(accountId, request);
     }
-
-    public ReservationSlot getReservationSlotById(String slotId){
-        return reservationSlotRepository.findById(slotId)
-                .orElseThrow(() -> {
-                    log.warn("예약 슬롯을 찾을 수 없습니다. slotId={}", slotId);
-                    return new ReservationException(ReservationErrorCode.RESERVATION_SLOT_NOT_FOUND);
-                });
-    }
-
-    public Account getAccountById(String accountId){
-        return accountRepository.findById(accountId)
-                .orElseThrow(() -> {
-                    log.warn("계정을 찾을 수 없습니다. accountId={}", accountId);
-                    return new ReservationException(ReservationErrorCode.RESERVATION_NOT_FOUND);
-                });
-    }
-
-    public ReservationHistory getHistoryById(String historyId){
-        return reservationHistoryRepository.findById(historyId)
-                .orElseThrow(() -> {
-                    log.warn("예약을 찾을 수 없습니다. historyId={}", historyId);
-                    return new ReservationException(ReservationErrorCode.RESERVATION_NOT_FOUND);
-                });
-    }
-
 
 }
